requirements.txt
pygame==2.5.2
openai==1.3.7
python-dotenv==1.0.0
pillow==10.1.0
requests==2.31.0
